+++ src/app/[locale]/(main)/(trip)/_components/UnsavedChangesWarning.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import {ReactNode, useCallback, useEffect} from 'react';
+import {ReactNode, useEffect} from 'react';
 import {useRouter} from '@/i18n/routing';
 
 export default function UnsavedChangesWarning({
@@ -9,18 +9,6 @@ export default function UnsavedChangesWarning({
   children: ReactNode;
 }) {
   const router = useRouter();
-
-  const handlePopState = useCallback(() => {
-    history.pushState(null, '', location.href);
-
-    const confirmLeave = window.confirm(
-      '페이지를 나가시겠습니까? 저장되지 않은 변경사항이 있을 수 있습니다.',
-    );
-
-    if (confirmLeave) {
-      router.push('/');
-    }
-  }, [router]);
 
   useEffect(() => {
     const handleBeforeUnload = (e: BeforeUnloadEvent) => {
@@ -43,21 +31,13 @@ export default function UnsavedChangesWarning({
       router.push('/');
     }
 
-    history.pushState(null, '', location.href);
-    window.addEventListener('popstate', handlePopState);
-
     sessionStorage.setItem('pageLoadTimestamp', currentTime);
-
-    window.addEventListener('beforeunload', handleBeforeUnload);
-    window.addEventListener('pageshow', handlePageShow);
 
     return () => {
       window.removeEventListener('beforeunload', handleBeforeUnload);
       window.removeEventListener('pageshow', handlePageShow);
-      window.removeEventListener('popstate', handlePopState);
-      sessionStorage.removeItem('pageLoadTimestamp');
     };
-  }, [router, handlePopState]);
+  }, [router]);
 
   return <>{children}</>;
 }
