--- src/app/[locale]/(main)/(trip)/time/page.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import React from 'react';
-
-import TripTimeConfigFrom from '@/app/[locale]/(main)/(trip)/_components/TripTimeConfigFrom';
-import TripTimeConfigHeader from '@/app/[locale]/(main)/(trip)/_components/TripTimeConfigHeader';
-import NotificationBox from '@/app/[locale]/(main)/(trip)/_components/NotificationBox';
-
-export default function TimePage() {
-  return (
-    <>
-      <TripTimeConfigHeader>
-        <NotificationBox />
-      </TripTimeConfigHeader>
-      <TripTimeConfigFrom />
-    </>
-  );
-}
