+++ src/app/[locale]/(main)/(trip)/trip/_components/NotificationBox.tsx
