+++ src/app/[locale]/(main)/(trip)/trip/_components/TripTimeConfigHeader.tsx
