+++ src/app/[locale]/(main)/(trip)/trip/edit/page.tsx
