+++ src/app/[locale]/(main)/(trip)/trip/layout.tsx
@@ -1,5 +1,4 @@
 import {ReactNode} from 'react';
-
 import UnsavedChangesWarning from '@/app/[locale]/(main)/(trip)/_components/UnsavedChangesWarning';
 
 export default function TripLayout({children}: {children: ReactNode}) {
