+++ src/app/[locale]/(main)/(trip)/trip/place/add/page.tsx
