+++ src/app/[locale]/(main)/(trip)/trip/ready/page.tsx
