+++ src/app/[locale]/(main)/layout.tsx
@@ -1,7 +1,7 @@
 import React, {ReactNode} from 'react';
 
-import Header from '@/components/Header';
 import Footer from '@/components/Footer';
+import Header from '@/components/Header';
 
 interface Props {
   children: ReactNode;
